rewrite printf wit recasts to accomodate fussy Pi compilers

+++ src-lib/oo-logprims.c
@@ -332,8 +332,8 @@ void
 
 if (ctx->verbosity > 9)
 {
-  fprintf(stderr, "DEBUG: penab %d olen %ld ilen %ld\n",
-    print_enable, strlen(trace_out_buffer), strlen(trace_in_buffer));
+  fprintf(stderr, "DEBUG: penab %d olen %d ilen %d\n",
+    print_enable, (int)strlen(trace_out_buffer), (int)strlen(trace_in_buffer));
 }
 
   clock_gettime (CLOCK_REALTIME, &current_time_fine);
